Update the way the builtin git completions are loaded
As noted in https://fishshell.com/docs/current/relnotes.html#changes-to-self-installing-builds, fish does not store its
data on the filesystem anymore, instead embedding it in the binary. This means that various fish paths no longer include
fish data paths. In order to properly load the original git completions, we source the file from fish using
`status get-file` instead of trying to find it from the filesystem.